workers/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "flipdisc-workers"
version = "0.1.0"
description = "Flip-Disc workers (Python)"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.24",
  "Pillow>=10.0",
]

[tool.setuptools]
# Workers are executed directly from source; we don't ship importable packages.
# This avoids backend errors like "Multiple top-level packages discovered".
packages = []
